// src/core/common/enums/account.enum.ts
export enum SchedulePeriod {
  DAILY = 'DAILY',
  WEEKLY = 'WEEKLY',
  TWO_WEEKS = 'TWO_WEEKS',
  MONTHLY = 'MONTHLY',
}

export enum ScheduleExecuteDay {
  EVERY_MONDAY = 'EVERY_MONDAY',
  EVERY_TUESDAY = 'EVERY_TUESDAY',
  EVERY_WEDNESDAY = 'EVERY_WEDNESDAY',
  EVERY_THURSDAY = 'EVERY_THURSDAY',
  EVERY_FRIDAY = 'EVERY_FRIDAY',
  EVERY_SATURDAY = 'EVERY_SATURDAY',
  EVERY_SUNDAY = 'EVERY_SUNDAY',
  EVERY_FIRST_AND_FIFTEENTH_OF_MONTH = 'EVERY_FIRST_AND_FIFTEENTH_OF_MONTH',
  EVERY_FIRST_DAY_OF_MONTH = 'EVERY_FIRST_DAY_OF_MONTH',
  EVERY_LAST_DAY_OF_MONTH = 'EVERY_LAST_DAY_OF_MONTH',
}

export enum NotificationProvider {
  EMAIL = 'EMAIL',
  WHATSAPP = 'WHATSAPP',
  PUSH_NOTIFICATION = 'PUSH_NOTIFICATION',
}